backend/requirements.txt
fastapi
uvicorn
google-generativeai
PyPDF2
python-multipart
python-dotenv
